requirements.txt
# ── Core Server ──
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1

# ── LLM & Agent Framework ──
google-genai==1.0.0
langgraph==0.2.60

# ── Database (Supabase — cloud PostgreSQL) ──
supabase==2.15.0

# ── HTTP Client (async API calls) ──
aiohttp>=3.11.0

# ── PDF Report Generation ──
reportlab==4.2.0

# ── Tavily Search ──
tavily-python==0.5.0

# ── Utilities ──
python-multipart==0.0.9

# ── Test Dependencies ──
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
hypothesis>=6.100.0
aioresponses>=0.7.6
